# packages/contracts/openapi/curriculum.yaml
openapi: 3.1.0
info:
  title: Question Bank — Curriculum API
  version: 1.0.0
  description: |
    The curriculum context's HTTP contract and the source of truth for client
    types (BACKEND-ARCHITECTURE §3). Every operation names the command or query
    handler it delegates to in `x-handler`, which is what the F15 fitness check
    reconciles against the handler registry.

    JSON is camelCase throughout; the database's snake_case never reaches here.
servers:
  - url: https://api.questionbank.example/v1
    description: Production
...
paths:
  /v1/taxonomy-versions:
    post:
      operationId: createTaxonomyDraft
      x-handler: CreateTaxonomyDraft
      summary: Create a taxonomy draft
      tags: [taxonomy]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTaxonomyDraftRequest' }
      responses:
        '201':
          description: The draft was created.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaxonomyWriteResult' }
        '400': { $ref: '#/components/responses/ValidationProblem' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
    get:
      operationId: listTaxonomyVersions
      x-handler: ListTaxonomyVersions
      summary: List taxonomy versions of an exam family
      tags: [taxonomy]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/TaxonomyVersionSort'
        - $ref: '#/components/parameters/TaxonomyVersionFilterState'
        - name: examFamily
          in: query
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: A page of taxonomy versions.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaxonomyVersionPage' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
  /v1/taxonomy-versions/{taxonomyVersionId}:
    parameters:
      - $ref: '#/components/parameters/TaxonomyVersionId'
    get:
      operationId: getTaxonomyVersion
      x-handler: GetTaxonomyVersion
      summary: Read one taxonomy version
      tags: [taxonomy]
      responses:
        '200':
          description: The taxonomy version.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaxonomyVersionDetail' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
  /v1/taxonomy-versions/{taxonomyVersionId}/concept-nodes:
    parameters:
      - $ref: '#/components/parameters/TaxonomyVersionId'
    post:
      operationId: addConceptNode
      x-handler: AddConceptNode
      summary: Add a concept node to a draft
      tags: [taxonomy]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddConceptNodeRequest' }
      responses:
        '201':
          description: The node was added.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaxonomyWriteResult' }
        '400': { $ref: '#/components/responses/ValidationProblem' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
        '422': { $ref: '#/components/responses/RuleViolationProblem' }
  /v1/taxonomy-versions/{taxonomyVersionId}/concept-nodes/{conceptNodeId}:
    parameters:
      - $ref: '#/components/parameters/TaxonomyVersionId'
      - $ref: '#/components/parameters/ConceptNodeId'
    patch:
      operationId: moveConceptNode
      x-handler: MoveConceptNode
      summary: Move a concept node under a new parent
      tags: [taxonomy]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MoveConceptNodeRequest' }
      responses:
        '200':
          description: The node was moved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaxonomyWriteResult' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
        '422': { $ref: '#/components/responses/RuleViolationProblem' }
    delete:
      operationId: removeConceptNode
      x-handler: RemoveConceptNode
      summary: Remove a concept node from a draft
      tags: [taxonomy]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: The node was removed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaxonomyWriteResult' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
        '422': { $ref: '#/components/responses/RuleViolationProblem' }
  /v1/taxonomy-versions/{taxonomyVersionId}/prerequisite-edges:
    parameters:
      - $ref: '#/components/parameters/TaxonomyVersionId'
    post:
      operationId: addPrerequisiteEdge
      x-handler: AddPrerequisiteEdge
      summary: Add a prerequisite edge to a draft
      tags: [taxonomy]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddPrerequisiteEdgeRequest' }
      responses:
        '201':
          description: The edge was added.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaxonomyWriteResult' }
        '400': { $ref: '#/components/responses/ValidationProblem' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
        '422': { $ref: '#/components/responses/RuleViolationProblem' }
  /v1/taxonomy-versions/{taxonomyVersionId}/publication:
    parameters:
      - $ref: '#/components/parameters/TaxonomyVersionId'
    post:
      operationId: publishTaxonomyVersion
      x-handler: PublishTaxonomyVersion
      summary: Publish a taxonomy version
      tags: [taxonomy]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: The version was published.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaxonomyWriteResult' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
        '422': { $ref: '#/components/responses/RuleViolationProblem' }
  /v1/taxonomy-versions/{taxonomyVersionId}/concept-subtree:
    parameters:
      - $ref: '#/components/parameters/TaxonomyVersionId'
    get:
      operationId: getConceptSubtree
      x-handler: GetConceptSubtree
      summary: Read a concept subtree, optionally depth-limited
      tags: [taxonomy]
      parameters:
        - name: rootNodeId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: depthLimit
          in: query
          required: false
          schema: { type: integer, minimum: 0, maximum: 20 }
      responses:
        '200':
          description: The subtree.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConceptSubtree' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
  /v1/taxonomy-versions/{taxonomyVersionId}/concepts/{conceptIdentityId}/prerequisites:
    parameters:
      - $ref: '#/components/parameters/TaxonomyVersionId'
      - name: conceptIdentityId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getConceptPrerequisites
      x-handler: GetConceptPrerequisites
      summary: Read what a concept requires and what requires it
      tags: [taxonomy]
      responses:
        '200':
          description: The prerequisite relations.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConceptPrerequisites' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
  /v1/exams:
    post:
      operationId: createExam
      x-handler: CreateExam
      summary: Create an exam
      tags: [exam-profiles]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateExamRequest' }
      responses:
        '201':
          description: The exam was created.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [examId]
                properties:
                  examId: { type: string, format: uuid }
        '400': { $ref: '#/components/responses/ValidationProblem' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
    get:
      operationId: listExams
      x-handler: ListExams
      summary: List exams
      tags: [exam-profiles]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/ExamSort'
        - $ref: '#/components/parameters/ExamFilterJurisdiction'
      responses:
        '200':
          description: A page of exams.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamPage' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
  /v1/exam-profile-versions:
    post:
      operationId: createProfileDraft
      x-handler: CreateProfileDraft
      summary: Create an exam profile draft
      tags: [exam-profiles]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProfileDraftRequest' }
      responses:
        '201':
          description: The draft was created.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileWriteResult' }
        '400': { $ref: '#/components/responses/ValidationProblem' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
        '422': { $ref: '#/components/responses/RuleViolationProblem' }
  /v1/exam-profile-versions/{profileVersionId}:
    parameters:
      - $ref: '#/components/parameters/ProfileVersionId'
    get:
      operationId: getExamProfileVersion
      x-handler: GetExamProfileVersion
      summary: Read one exam profile version
      tags: [exam-profiles]
      responses:
        '200':
          description: The profile version.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamProfileVersionDetail' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
    put:
      operationId: updateProfileDraft
      x-handler: UpdateProfileDraft
      summary: Replace the content of a profile draft
      tags: [exam-profiles]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileDraftContent' }
      responses:
        '200':
          description: The draft was updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileWriteResult' }
        '400': { $ref: '#/components/responses/ValidationProblem' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
        '422': { $ref: '#/components/responses/RuleViolationProblem' }
  /v1/exam-profile-versions/{profileVersionId}/publication:
    parameters:
      - $ref: '#/components/parameters/ProfileVersionId'
    post:
      operationId: publishProfileVersion
      x-handler: PublishProfileVersion
      summary: Publish an exam profile version
      tags: [exam-profiles]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                activate: { type: boolean, default: false }
      responses:
        '200':
          description: The profile version was published.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileWriteResult' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
        '422': { $ref: '#/components/responses/RuleViolationProblem' }
  /v1/exam-profile-versions/{profileVersionId}/supersession:
    parameters:
      - $ref: '#/components/parameters/ProfileVersionId'
    post:
      operationId: supersedeProfileVersion
      x-handler: SupersedeProfileVersion
      summary: Supersede a published profile version
      tags: [exam-profiles]
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: The profile version was superseded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileWriteResult' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
        '422': { $ref: '#/components/responses/RuleViolationProblem' }
  /v1/taxonomy-migrations:
    post:
      operationId: createMigration
      x-handler: CreateMigration
      summary: Create a taxonomy migration
      tags: [migrations]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMigrationRequest' }
      responses:
        '201':
          description: The migration was created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MigrationWriteResult' }
        '400': { $ref: '#/components/responses/ValidationProblem' }
        '403': { $ref: '#/components/responses/AuthorizationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
...
components:
  headers:
    ETag:
      description: The aggregate version, quoted. Send it back as If-Match.
      schema: { type: string }
    CorrelationId:
      description: The correlation id, echoed on every response including errors.
      schema: { type: string }
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema: { type: string, maxLength: 64 }
...
  responses:
    ValidationProblem:
      description: The request was malformed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    AuthorizationProblem:
      description: The principal may not perform this operation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  schemas:
...
    TaxonomyWriteResult:
      type: object
      additionalProperties: false
      required: [taxonomyVersionId, aggregateVersion]
      properties:
        taxonomyVersionId: { type: string, format: uuid }
        aggregateVersion: { type: integer, minimum: 1 }
...
    CreateTaxonomyDraftRequest:
      type: object
      additionalProperties: false
      required: [examFamily, academicYear]
      properties:
        examFamily: { type: string, minLength: 1 }
        academicYear: { type: string, pattern: '^\d{4}(-\d{2})?$' }
...